Add files via upload
Serializer just does the opposite of what the deserializer does... surprise.  CLKS is the test bench and has a data_i input that is 16 bits representing the inputted word from memory that it will read. run the TB for about 40 mircroseconds and you should see output.  Serializer uses the same Clock_Scalar already in the file repository.